requirements.txt
pycryptodomex
pywin32